+++ packages/registration_delivery/lib/pages/beneficiary_registration/beneficiary_acknowledgement.dart
@@ -1,11 +1,14 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:collection/collection.dart';
 import 'package:digit_components/digit_components.dart';
+import 'package:digit_data_model/models/entities/identifier_types.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../utils/i18_key_constants.dart' as i18;
 import '../../../widgets/localized.dart';
 import '../../blocs/search_households/search_bloc_common_wrapper.dart';
+import '../../blocs/search_households/search_households.dart';
 import '../../router/registration_delivery_router.gm.dart';
 
 @RoutePage()
@@ -25,6 +28,15 @@ class BeneficiaryAcknowledgementPage extends LocalizedStatefulWidget {
 
 class BeneficiaryAcknowledgementPageState
     extends LocalizedState<BeneficiaryAcknowledgementPage> {
+  late final HouseholdMemberWrapper? wrapper;
+
+  @override
+  void initState() {
+    super.initState();
+    final bloc = context.read<SearchHouseholdsBloc>();
+    wrapper = bloc.state.householdMembers.lastOrNull;
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -50,9 +62,22 @@ class BeneficiaryAcknowledgementPageState
         description: localizations.translate(
           i18.acknowledgementSuccess.acknowledgementDescriptionText,
         ),
+        subLabel: getSubText(wrapper),
         label: localizations
             .translate(i18.acknowledgementSuccess.acknowledgementLabelText),
       ),
     );
+  }
+
+  getSubText(HouseholdMemberWrapper? wrapper) {
+    return wrapper != null
+        ? '${localizations.translate(i18.beneficiaryDetails.beneficiaryId)}\n'
+            '${wrapper.members?.lastOrNull!.name!.givenName} - '
+            '${wrapper.members?.lastOrNull!.identifiers!.lastWhereOrNull(
+                  (e) =>
+                      e.identifierType ==
+                      IdentifierTypes.uniqueBeneficiaryID.toValue(),
+                )?.identifierId ?? localizations.translate(i18.common.noResultsFound)}'
+        : '';
   }
 }